fix(mdm): align protection lease duration contract
